adds an argument parser to the server
by adding an argument parser, we are able to easily allow the user to
distinguish various parts of their message (i.e "user",
"topic/channel", "message"). They can give us all, or none of the
aforementioned message parts which we can use to dynamically construct
a tuple pattern to match on for GET requests. PUT requests remain
unchanged, and should have a user, topic/channel and message to be a
complete, well-formatted request.