a kind of working first pipeline
Not sure what happened, but the code is working for some situations.
There are still a lot of bugs involved with electing the leader, for
example, sometimes we get into a deadlock state where everyone becomes
a candidate at the same time and refuses to vote for anyone else. Not
sure how to go about resolving this just yet, as I thought the
randomized election timeouts were supposed to stagger the nodes enough
so that new leaders are elected rather quickly.

Going to look more into it, but for now this (under some rare and hard
to reproduce cirumstances) allows us to write and read from a
connected tuplespace process group. We still need to introduce the
notion of a commit log, and then distribute that to all the raft nodes
in the cluster configuration, however, we have proven that reading and
writing to tuplespaces over ZMQ sockets is feasible.